uses active instances query directly
